fix/fix home page design
