+++ frontend/src/components/home/home-skeleton.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from './ui/skeleton';
+import { Skeleton } from '../ui/skeleton';
 
 const HomeSkeleton = () => {
   return (
